pyproject.toml
[project]
name = "tech_test_hoortrade"
version = "0.1.0"
description = "A Python package for Hoortrade technical test"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "black>=25.1.0",
    "coverage>=7.9.1",
    "django>=5.2.3",
    "django-cors-headers>=4.7.0",
    "django-environ>=0.12.0",
    "django-filter>=25.1",
    "django-polymorphic>=4.1.0",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "isort>=6.0.1",
    "markdown>=3.8.1",
    "pillow>=11.2.1",
    "pre-commit>=4.2.0",
    "psycopg>=3.2.9",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pytest-django>=4.11.1",
]

[tool.black]
line-length = 110
exclude = '''
/(
    \.git
  | __pycache__
  | old
  | build
  | dist
  | venv
  | frontend
)/
'''
[tool.isort]
profile = "black"
skip_glob = ["*/migrations/*"]
